requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
-e git+http://gitlab.hy01.internal.nosa.me/sre-team/kazoo.git#egg=kazoo
-e git+http://gitlab.hy01.internal.nosa.me/sre-team/python-lazy-object-proxy.git#egg=lazy-object-proxy
backports.ssl-match-hostname==3.4.0.2  # via tornado
blinker==1.4
certifi==2015.9.6.2       # via tornado
enum34==1.0.4
ez-setup==0.9
gevent==1.1b2
google-apputils==0.4.2
greenlet==0.4.9           # via gevent
ipaddress==1.0.14
jinja2==2.8
m3-cdecimal==2.3
markupsafe==0.23          # via jinja2
meld3==1.0.2              # via supervisor
mistune==0.7.1
msgpack-python==0.4.6
protobuf==2.6.1
pycrypto==2.6.1
pynginxconfig==0.3.4
python-dateutil==2.4.2    # via google-apputils
python-gflags==2.0        # via google-apputils
pytz==2015.6              # via google-apputils
pyyaml==3.11
raven==5.5.0
redis==2.10.3
simplejson==3.8.0
six==1.10.0               # via python-dateutil, sqlalchemy-utils
sqlalchemy-utils==0.31.0
sqlalchemy==1.0.8
supervisor==3.1.3
tornado==4.2.1
ujson==1.33
uwsgi==2.0.11.2
werkzeug==0.10.4

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools==18.4          # via protobuf
